+++ noughts-and-crosses/src/main/java/kata/os/and/xs/interfaces/OsAndXsAgent.java
@@ -3,5 +3,5 @@ package kata.os.and.xs.interfaces;
 import kata.game.framework.models.PlayPiece;
 import kata.game.framework.service.IGameAgent;
 
-public interface OsAndXsAgent extends IGameAgent<Player, Integer, PlayPiece<Player>> {
+public interface OsAndXsAgent extends IGameAgent<Player, Double, PlayPiece<Player>> {
 }
